+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Added `NEPTUNE_MODE` environment variable ([#928](https://github.com/neptune-ai/neptune-client/pull/928))
 - Added support of Service account management ([#927](https://github.com/neptune-ai/neptune-client/pull/927))
 
+## Fixes
+- More strict runtime requirements for matplotlib due to plotly incompatibility ([#960](https://github.com/neptune-ai/neptune-client/pull/960))
+
 ## neptune-client 0.16.4
 
 ## Fixes